api/requirements.txt
amqp==5.1.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
flower==1.0.0
humanize==4.0.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
tornado==6.1
urllib3==1.26.8
vine==5.0.0
watchdog==2.1.5
wcwidth==0.2.5
websocket-client==1.3.1
Werkzeug==2.0.3
google-api-python-client==2.50.0
httplib2==0.20.4
oauth2client==4.1.3
